+++ libtransmission/utils.h
@@ -238,7 +238,7 @@ uint64_t       tr_date( void );
 void           tr_wait( uint64_t delay_milliseconds );
 
 char*          tr_utf8clean( const char  * str,
-                             ssize_t       max_len,
+                             int           max_len,
                              tr_bool     * err );
